replace AuthActions with AuthActions2
